前端改为FIR降采样 加入信道选择滤波

+++ cpp_ver/demod.h
@@ -31,6 +31,6 @@ extern Lowpass<double> lowpassBaud;
 extern MovingAverageUtil<double, double, 2048> preambleMovingAverage;
 extern double magsqRaw;
 
-void processOneSample(int8_t i, int8_t q);
+void processOneSample(float i, float q);
 
 #endif
